pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "predictive-maintenance-rul"
version = "0.1.0"
description = "Portfolio template: FastAPI + Streamlit + supervised ML scaffold."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Mohamad Khoja" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "joblib>=1.4",
  "pydantic>=2.7",
  "requests>=2.32",
  "pandera>=0.21",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
  "tabulate>=0.9",

]

[project.optional-dependencies]
api = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
]
app = [
  "streamlit>=1.36",
]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5.6",
  "mypy>=1.10",
  "httpx>=0.27",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --disable-warnings --maxfail=1"
